Hire-Hawk/Hire-Hawk/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
Django==4.2.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.2
packaging==23.2
preshed==3.0.9
pydantic==2.5.2
pydantic_core==2.14.5
PyPDF2==3.0.1
python-docx==1.1.0
regex==2023.10.3
requests==2.31.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
thinc==8.2.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
